requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.2.0
gradio>=3.0.0
numpy>=1.19.5